Cleaned up Sway config after sway 1.0 update
With this commit I cleaned up the sway config as it was broken after the
update to version 1.0. I also added a new binding for locking hte
screen. With `swaylock` no longer part of `sway`, I installed
`swaylock-blur` for locking my screen, with the binding `$mod-x`.
I removed no longer needed code that was used for locking by Tronje.