First bare install script
With this commit I started working on an install script to make
installing arch on a machine easier. Not that I would do it that often
or would need the help. It is more of a list where I can note down which
packages I have installed. Because this is what i struggle most with
when installing Arch, remenbering all the packages you need or want.