Nvim plugs installation / upadate alias
In this commit I added an alias to update my packages as well as my nvim
plugs in one command. Not that this command would really be necessary,
but otherwise I'll be forgetting to update my nvim plugs regularly. So
here we go.

In addition, I also added a line to the install script that
automatically installs the nvim plugs after the config has been set in
place.