Deoplete fix
In this commit I fixed my deoplete config as it broke with the last
update, throwing the error that the function
`deoplete#mappings#manual_complete()` was not found.
As I don't really have an idea what I am doing I just copied code from
the internet and it worked half way. I removed the parts that didn't and
ended up with this.

In addition I added functionality for tabbing backwards through the
completions. Nice. But as this `delimitMate` plugin was interfering I
removed it and I think I am happier now with the way backets and
quotation marks are autoclosed. Lets see if this holds up.

Furthermore, I added a package that is required as a dependecy for
deoplete. Not really sure what it does, it seemingly didn't make a
difference in the functionality of deoplete, but oh well...