Select Display server on login
In this commit I added xorg configuration, mostly copied it from
upstream. With this I now can run xorg as display server, instead of
wayland. This is nice, because uder xorg, skype does not crash when
trying to screen share.

But I prefer to use Wayland, so I added a small propmt when logging in
that asks me which display server to use. Probably really stupid and one
wouldn't really need it, but hey, I like fumbling around.

So now in the `.zlogin` I have a small piece of code that asks the user
which display server to use and then pieces together the config for the
window manager and then starts the server.