Updated install script
In this commit I updated my install script to work on osx. Also, I
adapted some paths in my '.zshrc' to match the osx file structure. Then
I also disabled the ls command alias, as ls is working slightly
different on darwin. Connected to this, I also added a switch that
detects darwin use and enabled dircolors in a different way that on
GNU-Linux.