Real U.S. International Layout
In this commit I took the time to recreate the U.S. International layout
just the way I know it from Windows and Linux.

This means I changed all the Alt-GR and Shift+Alt-GR keys to the way
they are arranged on Linux and Windows. I also stripped all other mappings,
so that blank keys on the U.S. International layout as found on
Wikipedia are blank on this one as well. Just one mapping I left: The
one for the Apple logo.

This will spare me a lot of nerves when programming or entering
Passwords and ocasionally when I write in German and need some nice
Umlate.