Removed display server selection
I built this in, when I was still mainly on Arch and needed a display
server that would allow me to share my screen. I could not get it to
work with Wayland, so I installed Xorg as well. But as I mostly wanted
to use Wayland, I implemented this siwtch.

Totally unnecessary for Mac.

Also I don't need to start my SSH agent on Mac.